+++ src/defaults.h
@@ -1,4 +1,5 @@
 #define Sprintf(f, ...) ({ char* s; asprintf(&s, f, __VA_ARGS__); String r = s; free(s); r; })
+#define Stdprintf(f, ...) ({ char* s; asprintf(&s, f, __VA_ARGS__); std::string r = s; free(s); r; })
 #define ESPMAC (Sprintf("%06" PRIx64, ESP.getEfuseMac() >> 24))
 
 #ifdef VERBOSE