+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3]
+
+- Don't Query until device rssi > -90dBm
+- Query forever w/ exponential backoff (retry after maxes at 1m)
+- New name fingerprint for macbooks
+- Query: use max power for all connections
+
 ## [2.0.2]
 
 - Don't Query until device is <= maxDistance